interview-coach/server/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
pgvector==0.3.5
redis==5.0.8
httpx==0.27.2
langchain==0.3.7
langgraph==0.2.44
langsmith==0.1.147
openai==1.51.2
anthropic==0.39.0
tiktoken==0.7.0
numpy==2.1.3
tenacity==9.0.0
email-validator==2.2.0
jinja2==3.1.4
pytest==8.3.3
anyio==4.6.2.post1
temporalio==1.8.0